factor out ModuleSpecifier to //core/module_specifier.rs
