+++ core/lib.rs
@@ -8,6 +8,7 @@ mod flags;
 mod isolate;
 mod js_errors;
 mod libdeno;
+mod module_specifier;
 mod modules;
 mod shared_queue;
 
@@ -16,6 +17,7 @@ pub use crate::isolate::*;
 pub use crate::js_errors::*;
 pub use crate::libdeno::deno_mod;
 pub use crate::libdeno::PinnedBuf;
+pub use crate::module_specifier::ModuleSpecifier;
 pub use crate::modules::*;
 
 pub fn v8_version() -> &'static str {
